Move `ProductionCost` to `libOPHD`
