+++ appOPHD/MapObjects/Structures/Factory.cpp
@@ -1,7 +1,8 @@
 #include "Factory.h"
 
-#include "../../ProductionCost.h"
 #include "../../States/MapViewStateHelper.h" // yuck
+
+#include <libOPHD/ProductionCost.h>
 
 #include <NAS2D/Dictionary.h>
