+++ appOPHD/UI/Reports/FactoryReport.cpp
@@ -8,12 +8,13 @@
 #include "../../Constants/Strings.h"
 #include "../../Constants/UiConstants.h"
 #include "../../StructureManager.h"
-#include "../../ProductionCost.h"
 #include "../../ProductCatalogue.h"
 
 #include "../../MapObjects/Structures/SurfaceFactory.h"
 #include "../../MapObjects/Structures/SeedFactory.h"
 #include "../../MapObjects/Structures/UndergroundFactory.h"
+
+#include <libOPHD/ProductionCost.h>
 
 #include <NAS2D/Utility.h>
 #include <NAS2D/Renderer/Renderer.h>
